--- assignment1/src/utils/sprite.js
@@ -1,12 +1,0 @@
-class Sprite {
-    constructor(x, y, width, height, src) {
-        this.X = x;
-        this.Y = y;
-        this.image = new Image();
-        this.image.width = width;
-        this.image.height = height;
-        this.image.src = src;
-    }
-}
-
-export default Sprite;
